WCAnalyzer - WIP

+++ WCAnalyzer/WCAnalyzer.Core/Models/Placement.cs
@@ -36,7 +36,12 @@ public abstract class Placement
     /// <summary>
     /// Gets or sets the flags of the placement.
     /// </summary>
-    public int Flags { get; set; }
+    /// <remarks>
+    /// In Warcraft.NET, this is a ushort that represents various flags:
+    /// For MDDFFlags: 0x40 = NameIdIsFiledataId (among other flags)
+    /// For MODFFlags: 0x8 = NameIdIsFiledataId (among other flags)
+    /// </remarks>
+    public ushort Flags { get; set; }
     
     /// <summary>
     /// Gets or sets the FileDataID of the placement.
@@ -60,6 +65,10 @@ public class ModelPlacement : Placement
     /// <summary>
     /// Gets or sets the scale of the model.
     /// </summary>
+    /// <remarks>
+    /// In Warcraft.NET MDDFEntry, the raw data is a ushort ScalingFactor where 1024 = 1.0f.
+    /// This property contains the converted float value after division by 1024.
+    /// </remarks>
     public float Scale { get; set; }
 }
 
@@ -81,5 +90,8 @@ public class WmoPlacement : Placement
     /// <summary>
     /// Gets or sets the scale of the WMO. Scale flag must be set to use this value.
     /// </summary>
+    /// <remarks>
+    /// In Warcraft.NET, when the MODFFlags.HasScale flag (0x4) is set, this value is used to scale the WMO.
+    /// </remarks>
     public ushort Scale { get; set; }
 }
